requirements.txt
Django~=3.2.15
django-debug-toolbar~=3.8.1
django-nested-inline~=0.4.5